+++ CHANGELOG.rst
@@ -6,6 +6,9 @@ CHANGELOG
 Unreleased
 ==========
 
+Changed
+-------
+* Expose fintrist.mongoclient.
 
 ==================
 1.0.0 - 2020-12-05