+++ CHANGELOG.rst
@@ -9,6 +9,9 @@ Unreleased
 Added
 -----
 * Can toggle into a Test database.
+* NNModel study type.
+* create_nn for creating NN Study objects.
+* learn.py to fintrist_lib, with NN manipulation functions.
 
 Changed
 -------