+++ CHANGELOG.rst
@@ -6,6 +6,10 @@ CHANGELOG
 Unreleased
 ==========
 
+==================
+1.1.0 - 2021-01-21
+==================
+
 Added
 -----
 * Can toggle into a Test database.